Replace the AtomApplication test suite
